some pglite

+++ packages/core/src/database/index.ts
@@ -157,6 +157,27 @@ export const createDatabase = async ({
 
     const kyselyDialect = new KyselyPGlite(driver.instance).dialect;
 
+    const role = "ponder";
+
+    // TODO(kyle) role might already be on "ponder";
+
+    await driver.instance.query(
+      `CREATE SCHEMA IF NOT EXISTS "${preBuild.namespace}"`,
+    );
+    const hasRole = await driver.instance
+      .query(`SELECT FROM pg_roles WHERE rolname = '${role}'`)
+      .then(({ rows }) => rows[0]);
+    if (hasRole) {
+      await driver.instance.query(`DROP OWNED BY ${role}`);
+      await driver.instance.query(`DROP ROLE IF EXISTS ${role}`);
+    }
+    await driver.instance.query(`CREATE ROLE ${role} WITH LOGIN`);
+    await driver.instance.query(`GRANT postgres TO ${role}`);
+    await driver.instance.query(`SET ROLE ${role}`);
+    await driver.instance.query(
+      `SET search_path TO ${preBuild.namespace}, public`,
+    );
+
     qb = {
       internal: new HeadlessKysely({
         name: "internal",
@@ -246,8 +267,13 @@ export const createDatabase = async ({
     const role = "ponder_readonly";
 
     await internal.query(`CREATE SCHEMA IF NOT EXISTS "${preBuild.namespace}"`);
-    await internal.query(`DROP OWNED BY ${role}`);
-    await internal.query(`DROP ROLE IF EXISTS ${role}`);
+    const hasRole = await internal
+      .query(`SELECT FROM pg_roles WHERE rolname = '${role}'`)
+      .then(({ rows }) => rows[0]);
+    if (hasRole) {
+      await internal.query(`DROP OWNED BY ${role}`);
+      await internal.query(`DROP ROLE IF EXISTS ${role}`);
+    }
     await internal.query(`CREATE ROLE ${role} WITH LOGIN`);
     await internal.query(
       `GRANT CONNECT ON DATABASE "${connection.database}" TO ${role}`,
